requirements.txt
Flask>=3.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
reportlab>=4.0.0
